+++ api/tests/authentication/integration/routes.test.js
@@ -1,6 +1,6 @@
 import querystring from 'node:querystring';
 
-import { tokenController } from '../../../src/authentication/application/token-controller.js';
+import { tokenController } from '../../../src/authentication/application/token.controller.js';
 import { createServer, expect, sinon } from '../../test-helper.js';
 
 describe('Integration | Authentication | Application | Router', function () {
